requirements.txt
Django>=4.2.0
pytest>=7.3.1
pytest-django>=4.5.2
djangorestframework==3.14.0
coreapi==2.3.3
PyYAML==6.0.1 
# drf-standardized-errors==0.14.1
django-cors-headers==4.6.0